backend/requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
sqlmodel==0.0.16
pydantic==2.6.1
python-multipart==0.0.9
pytest
httpx
numpy
scipy
passlib[bcrypt]
bcrypt==3.2.2
python-jose[cryptography]
python-multipart
